+++ MatrixSDK/Crypto/MXCrypto.m
@@ -507,24 +507,21 @@ NSTimeInterval kMXCryptoMinForceSessionPeriod = 3600.0; // one hour
 #endif
 }
 
-- (BOOL)hasKeysToDecryptEvent:(MXEvent *)event
+- (void)hasKeysToDecryptEvent:(MXEvent *)event onComplete:(void (^)(BOOL))onComplete
 {
-    __block BOOL hasKeys = NO;
-    
 #ifdef MX_CRYPTO
     
     // We need to go to decryptionQueue only to use getRoomDecryptor
     // Other subsequent calls are thread safe because of the implementation of MXCryptoStore
-    dispatch_sync(decryptionQueue, ^{
+    dispatch_async(decryptionQueue, ^{
         NSString *algorithm = event.wireContent[@"algorithm"];
         id<MXDecrypting> alg = [self getRoomDecryptor:event.roomId algorithm:algorithm];
         
-        hasKeys = [alg hasKeysToDecryptEvent:event];
+        BOOL hasKeys = [alg hasKeysToDecryptEvent:event];
+        onComplete(hasKeys);
     });
     
 #endif
-    
-    return hasKeys;
 }
 
 - (MXEventDecryptionResult *)decryptEvent:(MXEvent *)event inTimeline:(NSString*)timeline